model_audio/conf/requirements.txt
cython
Keras==2.4.3
Keras-Preprocessing==1.1.2
librosa==0.7.0
numba==0.48.0
numpy==1.20.1
pandas==1.2.0
pydub==0.24.1
PyYAML==5.3.1
urllib3==1.26
soundfile
gdown
google-api-core==1.26.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-cloud-speech==2.0.1
google-pasta==0.2.0
boto3==1.17.5
pygelf==0.4.0
psycopg2-binary==2.8.6
scikit-learn
pytz
